+++ test/pip_license_checker/pypi_test.clj
@@ -106,18 +106,6 @@
 
 ;; pypi/classifiers->license
 
-(def params-get-first-longest-vector
-  [[nil nil "Nil vectors"]
-   [[] nil "Empty vectors collection"]
-   [[[1] [1 2 3] [1 2]] [1 2 3] "Collection of vectors of ints"]
-   [[["a" "b" "c"]  ["a" "b"] ["a"]] ["a" "b" "c"] "Collection of vectors of strings"]])
-
-(deftest test-get-first-longest-vector
-  (testing "Get first longest vector in collection"
-    (doseq [[vectors expected description] params-get-first-longest-vector]
-      (testing description
-        (is (= expected (pypi/get-first-longest-vector vectors)))))))
-
 (def params-classifiers->license
   [[nil nil "No classifiers"]
    [[] nil "Empty classifiers"]
@@ -128,21 +116,16 @@
     "MIT License"
     "License found"]
    [["Operating System :: Unix"
-     "License :: OSI Approved :: BSD License"
-     "License :: OSI Approved :: MIT License"]
-    "BSD License"
-    "Get last license"]
+     "License :: OSI Approved :: GNU General Public License (GPL)"
+     "License :: OSI Approved :: GNU Library or Lesser General Public License (LGPL)"
+     "License :: OSI Approved :: Mozilla Public License 1.1 (MPL 1.1)"]
+    "GNU General Public License (GPL), GNU Library or Lesser General Public License (LGPL), Mozilla Public License 1.1 (MPL 1.1)"
+    "Get list of all licenses"]
    [["Operating System :: Unix"
      "License :: OSI Approved :: MIT License"
      "License :: OSI Approved"]
     "MIT License"
     "Get most detailed license"]
-   [["License :: OSI Approved"
-     "License :: "
-     "License :: OSI Approved :: MIT License :: MIT with details"
-     "License :: OSI Approved :: MIT License"]
-    "MIT with details"
-    "Get most detailed license - 2"]
    [["Operating System :: Unix"
      "License :: OSI Approved"]
     nil